+++ skema/skema-rs/mathml/src/lib.rs
@@ -4,4 +4,4 @@ pub mod ast;
 pub mod graph;
 pub mod parsing;
 pub mod normalization;
-pub mod expression_tree;
+pub mod expression;
